backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
mysqlclient==2.2.7
django-cors-headers==4.6.0
python-dotenv==1.0.1
